:recycle: Refactor specs generation with hook_for
