+++ lib/generators/rest_api_generator/spec/templates/rspec/nested_resource_controller_spec.rb.tt
