+++ lib/generators/rest_api_generator/spec/templates/rspec/resource_controller_spec.rb.tt
