+++ lib/generators/rest_api_generator/spec_generator.rb
@@ -1,0 +1,1 @@
+clas
